requirements.txt
pandas
numpy
matplotlib
seaborn
geopandas
shapely.geometry